refactor: moved stats API responces
